Converted float to double and branched

+++ lib/ACESlib.ODT_Common.h
@@ -16,8 +16,8 @@ using namespace std;
 
 
 // Target white and black points for cinema system tonescale
-const float CINEMA_WHITE = 48.0f;
-const float CINEMA_BLACK = pow10f(log10f(0.02f)); // CINEMA_WHITE / 2400. 
+const double CINEMA_WHITE = 48.0;
+const double CINEMA_BLACK = pow10f(log10f(0.02)); // CINEMA_WHITE / 2400. 
     // CINEMA_BLACK is defined in this roundabout manner in order to be exactly equal to 
     // the result returned by the cinema 48-nit ODT tonescale.
     // Though the min point of the tonescale is designed to return 0.02, the tonescale is 
@@ -31,66 +31,66 @@ const float CINEMA_BLACK = pow10f(log10f(0.02f)); // CINEMA_WHITE / 2400.
 
 
 // Gamma compensation factor
-const float DIM_SURROUND_GAMMA = 0.9811f;
+const double DIM_SURROUND_GAMMA = 0.9811;
 
 // Saturation compensation factor
-const float ODT_SAT_FACTOR = 0.93f;
-const array<array <float, 3>, 3> ODT_SAT_MAT = calc_sat_adjust_matrix( ODT_SAT_FACTOR, AP1_RGB2Y);
+const double ODT_SAT_FACTOR = 0.93;
+const array<array <double, 3>, 3> ODT_SAT_MAT = calc_sat_adjust_matrix( ODT_SAT_FACTOR, AP1_RGB2Y);
 
 
 
 
-const array<array <float, 3>, 3> D60_2_D65_CAT = calculate_cat_matrix( AP0.white, REC709_PRI.white);
+const array<array <double, 3>, 3> D60_2_D65_CAT = calculate_cat_matrix( AP0.white, REC709_PRI.white);
 
 
 
 
-float Y_2_linCV( float Y, float Ymax, float Ymin) 
+double Y_2_linCV( double Y, double Ymax, double Ymin) 
 {
   return (Y - Ymin) / (Ymax - Ymin);
 }
 
-float linCV_2_Y( float linCV, float Ymax, float Ymin) 
+double linCV_2_Y( double linCV, double Ymax, double Ymin) 
 {
   return linCV * (Ymax - Ymin) + Ymin;
 }
 
-array <float, 3> Y_2_linCV_f3(const array <float, 3> &Y, float Ymax, float Ymin)
+array <double, 3> Y_2_linCV_f3(const array <double, 3> &Y, double Ymax, double Ymin)
 {
-    array <float, 3> linCV;
+    array <double, 3> linCV;
     linCV[0] = Y_2_linCV( Y[0], Ymax, Ymin);
     linCV[1] = Y_2_linCV( Y[1], Ymax, Ymin);
     linCV[2] = Y_2_linCV( Y[2], Ymax, Ymin);
     return linCV;
 }
 
-array <float, 3> linCV_2_Y_f3(const array <float, 3> &linCV, float Ymax, float Ymin)
+array <double, 3> linCV_2_Y_f3(const array <double, 3> &linCV, double Ymax, double Ymin)
 {
-    array <float, 3> Y;
+    array <double, 3> Y;
     Y[0] = linCV_2_Y( linCV[0], Ymax, Ymin);
     Y[1] = linCV_2_Y( linCV[1], Ymax, Ymin);
     Y[2] = linCV_2_Y( linCV[2], Ymax, Ymin);
     return Y;
 }
 
-array <float, 3> darkSurround_to_dimSurround(const array <float, 3> &linearCV)
+array <double, 3> darkSurround_to_dimSurround(const array <double, 3> &linearCV)
 {
-  array <float, 3> XYZ = mult_f3_f44( linearCV, AP1_2_XYZ_MAT);
-  array <float, 3> xyY = XYZ_2_xyY(XYZ);
-  xyY[2] = clamp( xyY[2], 0.0f, HALF_POS_INF);
+  array <double, 3> XYZ = mult_f3_f44( linearCV, AP1_2_XYZ_MAT);
+  array <double, 3> xyY = XYZ_2_xyY(XYZ);
+  xyY[2] = clamp( xyY[2], 0.0, HALF_POS_INF);
   xyY[2] = powf( xyY[2], DIM_SURROUND_GAMMA);
   XYZ = xyY_2_XYZ(xyY);
 
   return mult_f3_f44( XYZ, XYZ_2_AP1_MAT);
 }
 
-array <float, 3> dimSurround_to_darkSurround(const array <float, 3> &linearCV)
+array <double, 3> dimSurround_to_darkSurround(const array <double, 3> &linearCV)
 {
-  array <float, 3> XYZ = mult_f3_f44( linearCV, AP1_2_XYZ_MAT);
+  array <double, 3> XYZ = mult_f3_f44( linearCV, AP1_2_XYZ_MAT);
 
-  array <float, 3> xyY = XYZ_2_xyY(XYZ);
-  xyY[2] = clamp( xyY[2], 0.0f, HALF_POS_INF);
-  xyY[2] = powf( xyY[2], 1.0f/DIM_SURROUND_GAMMA);
+  array <double, 3> xyY = XYZ_2_xyY(XYZ);
+  xyY[2] = clamp( xyY[2], 0.0, HALF_POS_INF);
+  xyY[2] = powf( xyY[2], 1.0/DIM_SURROUND_GAMMA);
   XYZ = xyY_2_XYZ(xyY);
 
   return mult_f3_f44( XYZ, XYZ_2_AP1_MAT);
@@ -102,54 +102,54 @@ array <float, 3> dimSurround_to_darkSurround(const array <float, 3> &linearCV)
 /* ---- Functions to compress highlights ---- */
 // allow for simulated white points without clipping
 
-float roll_white_fwd( 
-    float in,      // color value to adjust (white scaled to around 1.0)
-    float new_wht, // white adjustment (e.g. 0.9 for 10% darkening)
-    float width    // adjusted width (e.g. 0.25 for top quarter of the tone scale)
+double roll_white_fwd( 
+    double in,      // color value to adjust (white scaled to around 1.0)
+    double new_wht, // white adjustment (e.g. 0.9 for 10% darkening)
+    double width    // adjusted width (e.g. 0.25 for top quarter of the tone scale)
   )
 {
-    const float x0 = -1.0f;
-    const float x1 = x0 + width;
-    const float y0 = -new_wht;
-    const float y1 = x1;
-    const float m1 = (x1 - x0);
-    const float a = y0 - y1 + m1;
-    const float b = 2.0f * ( y1 - y0) - m1;
-    const float c = y0;
-    const float t = (-in - x0) / (x1 - x0);
-    float out = 0.0f;
-    if ( t < 0.0f)
+    const double x0 = -1.0;
+    const double x1 = x0 + width;
+    const double y0 = -new_wht;
+    const double y1 = x1;
+    const double m1 = (x1 - x0);
+    const double a = y0 - y1 + m1;
+    const double b = 2.0 * ( y1 - y0) - m1;
+    const double c = y0;
+    const double t = (-in - x0) / (x1 - x0);
+    double out = 0.0;
+    if ( t < 0.0)
         out = -(t * b + c);
-    else if ( t > 1.0f)
+    else if ( t > 1.0)
         out = in;
     else
         out = -(( t * a + b) * t + c);
     return out;
 }
 
-float roll_white_rev( 
-    float in,      // color value to adjust (white scaled to around 1.0)
-    float new_wht, // white adjustment (e.g. 0.9 for 10% darkening)
-    float width    // adjusted width (e.g. 0.25 for top quarter of the tone scale)
+double roll_white_rev( 
+    double in,      // color value to adjust (white scaled to around 1.0)
+    double new_wht, // white adjustment (e.g. 0.9 for 10% darkening)
+    double width    // adjusted width (e.g. 0.25 for top quarter of the tone scale)
   )
 {
-    const float x0 = -1.0f;
-    const float x1 = x0 + width;
-    const float y0 = -new_wht;
-    const float y1 = x1;
-    const float m1 = (x1 - x0);
-    const float a = y0 - y1 + m1;
-    const float b = 2.0f * ( y1 - y0) - m1;
-    float c = y0;
-    float out = 0.0f;
+    const double x0 = -1.0;
+    const double x1 = x0 + width;
+    const double y0 = -new_wht;
+    const double y1 = x1;
+    const double m1 = (x1 - x0);
+    const double a = y0 - y1 + m1;
+    const double b = 2.0 * ( y1 - y0) - m1;
+    double c = y0;
+    double out = 0.0;
     if ( -in < y0)
         out = -x0;
     else if ( -in > y1)
         out = in;
     else {
         c = c + in;
-        const float discrim = sqrtf( b * b - 4.0f * a * c);
-        const float t = ( 2.0f * c) / ( -discrim - b);
+        const double discrim = sqrtf( b * b - 4.0 * a * c);
+        const double t = ( 2.0 * c) / ( -discrim - b);
         out = -(( t * ( x1 - x0)) + x0);
     }
     return out;
